--- connect.go
@@ -1,8 +1,0 @@
-// Handle database connection details, loaded from separate configuration file
-// or environment variables.
-
-package main
-
-func connectMySQL(connDetails struct{}) {
-
-}
