+++ database/connect.go
@@ -1,0 +1,8 @@
+// Handle database connection details, loaded from separate configuration file
+// or environment variables.
+
+package main
+
+func connectMySQL(connDetails struct{}) {
+
+}
